requirements.txt
discord.py[voice]>=2.3.0
qwen-tts>=0.1.0
torch>=2.0.0
soundfile>=0.12.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
